requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
decorator==4.1.2
Django==1.11.7
django-cors-headers==2.1.0
django-extensions==1.9.8
djangorestframework==3.7.3
future==0.16.0
geocoder==1.33.0
idna==2.6
psycopg2==2.7.3.2
pytz==2017.3
ratelim==0.1.6
requests==2.18.4
six==1.11.0
typing==3.6.2
urllib3==1.22
